refactor : Lotto.kt의 LottoRank 계산 메서드 제거
